chore: move aws folder
